EY-4310 Flytt kodeverk routes etc til behandling fra trygdetid
