+++ apps/etterlatte-behandling/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
     implementation(project(":libs:etterlatte-migrering-model"))
     implementation(project(":libs:etterlatte-brev-model"))
     implementation(project(":libs:etterlatte-vilkaarsvurdering-model"))
+    implementation(project(":libs:etterlatte-trygdetid-model"))
 
     implementation(libs.cache.caffeine)
     implementation(libs.navfelles.tokenvalidationktor2) {
