+++ apps/etterlatte-behandling/src/main/kotlin/kodeverk/KodeverkRoute.kt
@@ -1,4 +1,4 @@
-package no.nav.etterlatte.trygdetid
+package no.nav.etterlatte.kodeverk
 
 import io.ktor.server.application.call
 import io.ktor.server.response.respond
@@ -8,7 +8,7 @@ import io.ktor.server.routing.route
 import no.nav.etterlatte.libs.ktor.token.brukerTokenInfo
 
 fun Route.kodeverk(kodeverkService: KodeverkService) {
-    route("api/trygdetid/kodeverk/land") {
+    route("api/kodeverk/land") {
         get {
             call.respond(kodeverkService.hentAlleLand(brukerTokenInfo))
         }
