+++ apps/etterlatte-behandling/src/main/resources/application.conf
@@ -62,3 +62,7 @@ krr.scope = ${?KRR_SCOPE}
 
 axsys.url = ${?AXSYS_URL}
 axsys.scope = ${?AXSYS_SCOPE}
+
+kodeverk.client.id = ${?KODEVERK_CLIENT_ID}
+kodeverk.resource.url = ${?KODEVERK_URL}
+applicationName = ${?NAIS_APP_NAME}
