+++ apps/etterlatte-behandling/src/test/kotlin/kodeverk/KodeverkServiceTest.kt
@@ -1,5 +1,6 @@
-package no.nav.etterlatte.trygdetid
+package no.nav.etterlatte.kodeverk
 
+import LandNormalisert
 import com.fasterxml.jackson.module.kotlin.readValue
 import io.mockk.coEvery
 import io.mockk.coVerify
@@ -7,10 +8,6 @@ import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
 import no.nav.etterlatte.ktor.token.simpleSaksbehandler
 import no.nav.etterlatte.libs.common.objectMapper
-import no.nav.etterlatte.trygdetid.klienter.Beskrivelse
-import no.nav.etterlatte.trygdetid.klienter.Betydning
-import no.nav.etterlatte.trygdetid.klienter.KodeverkKlient
-import no.nav.etterlatte.trygdetid.klienter.KodeverkResponse
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 
@@ -69,6 +66,5 @@ internal class KodeverkServiceTest {
         }
     }
 
-    private fun opprettLandkoderResponse(): KodeverkResponse =
-        objectMapper.readValue(javaClass.getResource("/kodeverk/landkoder.json")!!.readText())
+    private fun opprettLandkoderResponse(): KodeverkResponse = objectMapper.readValue(javaClass.getResource("/landkoder.json")!!.readText())
 }
