+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/revurderingsoversikt/sluttbehandlingUtland/LandRad.tsx
@@ -1,8 +1,8 @@
 import { Select } from '@navikt/ds-react'
-import { ILand } from '~shared/api/trygdetid'
 import styled from 'styled-components'
 import DokumenterForLand from '~components/behandling/revurderingsoversikt/sluttbehandlingUtland/DokumenterForLand'
 import { LandMedDokumenter } from '~shared/types/RevurderingInfo'
+import { ILand } from '~utils/kodeverk'
 
 const Flex = styled.div`
   display: flex;
