+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/soeknadsoversikt/familieforhold/Familieforhold.tsx
@@ -8,10 +8,11 @@ import { Personopplysninger } from '~shared/types/grunnlag'
 import styled from 'styled-components'
 import { RedigerFamilieforhold } from '~components/behandling/soeknadsoversikt/familieforhold/RedigerFamilieforhold'
 import { useApiCall } from '~shared/hooks/useApiCall'
-import { hentAlleLand, ILand } from '~shared/api/trygdetid'
 import { mapApiResult, Result } from '~shared/api/apiUtils'
 import Spinner from '~shared/Spinner'
 import { finnLandSomTekst } from '~components/person/personopplysninger/utils'
+import { hentAlleLand } from '~shared/api/behandling'
+import { ILand } from '~utils/kodeverk'
 
 export interface PropsFamilieforhold {
   behandling: IDetaljertBehandling
