+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/personopplysninger/Innflytting.tsx
@@ -2,8 +2,8 @@ import React, { ReactNode } from 'react'
 import { Personopplysning } from '~components/person/personopplysninger/Personopplysning'
 import { AirplaneIcon } from '@navikt/aksel-icons'
 import { Heading, Table } from '@navikt/ds-react'
-import { ILand } from '~shared/api/trygdetid'
 import { finnLandSomTekst } from '~components/person/personopplysninger/utils'
+import { ILand } from '~utils/kodeverk'
 
 interface InnflyttingDTO {
   fraflyttingsland?: string
