+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/personopplysninger/Personopplysninger.tsx
@@ -13,11 +13,11 @@ import { Foreldre } from '~components/person/personopplysninger/Foreldre'
 import { AvdoedesBarn } from '~components/person/personopplysninger/AvdoedesBarn'
 import { Sivilstatus } from '~components/person/personopplysninger/Sivilstatus'
 import { Innflytting } from '~components/person/personopplysninger/Innflytting'
-import { hentAlleLand } from '~shared/api/trygdetid'
 import { Utflytting } from '~components/person/personopplysninger/Utflytting'
 import { Vergemaal } from '~components/person/personopplysninger/Vergemaal'
 import { hentFamilieOpplysninger } from '~shared/api/pdltjenester'
 import styled from 'styled-components'
+import { hentAlleLand } from '~shared/api/behandling'
 
 export const Personopplysninger = ({
   sakResult,
