+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/personopplysninger/utils.ts
@@ -1,4 +1,4 @@
-import { ILand } from '~shared/api/trygdetid'
+import { ILand } from '~utils/kodeverk'
 
 export const finnLandSomTekst = (isoLandKode: string, landListe: ILand[]): string | undefined => {
   return landListe.find((val) => val.isoLandkode === isoLandKode)?.beskrivelse.tekst
