+++ apps/etterlatte-trygdetid/.nais/prod.yaml
@@ -85,10 +85,6 @@ spec:
       value: http://etterlatte-grunnlag
     - name: ETTERLATTE_GRUNNLAG_CLIENT_ID
       value: prod-gcp.etterlatte.etterlatte-grunnlag
-    - name: KODEVERK_URL
-      value: "http://kodeverk-api.team-rocket/api/v1/kodeverk"
-    - name: KODEVERK_CLIENT_ID
-      value: prod-gcp.team-rocket.kodeverk-api
   envFrom:
     - secret: my-application-unleash-api-token
   accessPolicy:
@@ -103,7 +99,5 @@ spec:
       rules:
         - application: etterlatte-behandling
         - application: etterlatte-grunnlag
-        - application: kodeverk-api
-          namespace: team-rocket
       external:
         - host: etterlatte-unleash-api.nav.cloud.nais.io
