+++ apps/etterlatte-trygdetid/src/main/kotlin/trygdetid/config/ApplicationContext.kt
@@ -6,7 +6,6 @@ import no.nav.etterlatte.libs.common.Miljoevariabler
 import no.nav.etterlatte.libs.database.ApplicationProperties
 import no.nav.etterlatte.libs.database.DataSourceBuilder
 import no.nav.etterlatte.libs.ktor.httpClient
-import no.nav.etterlatte.trygdetid.KodeverkService
 import no.nav.etterlatte.trygdetid.TrygdetidBeregningService
 import no.nav.etterlatte.trygdetid.TrygdetidRepository
 import no.nav.etterlatte.trygdetid.TrygdetidServiceImpl
@@ -14,7 +13,6 @@ import no.nav.etterlatte.trygdetid.avtale.AvtaleRepository
 import no.nav.etterlatte.trygdetid.avtale.AvtaleService
 import no.nav.etterlatte.trygdetid.klienter.BehandlingKlient
 import no.nav.etterlatte.trygdetid.klienter.GrunnlagKlient
-import no.nav.etterlatte.trygdetid.klienter.KodeverkKlient
 
 class ApplicationContext {
     val config: Config = ConfigFactory.load()
@@ -27,7 +25,6 @@ class ApplicationContext {
         )
     private val grunnlagKlient = GrunnlagKlient(config, httpClient())
     val behandlingKlient = BehandlingKlient(config, httpClient())
-    val kodeverkService = KodeverkService(KodeverkKlient(config, httpClient()))
     val trygdetidService =
         TrygdetidServiceImpl(
             TrygdetidRepository(dataSource),
