+++ apps/etterlatte-trygdetid/src/main/resources/application.conf
@@ -18,10 +18,6 @@ behandling.resource.url = ${?ETTERLATTE_BEHANDLING_URL}
 grunnlag.client.id = ${?ETTERLATTE_GRUNNLAG_CLIENT_ID}
 grunnlag.resource.url = ${?ETTERLATTE_GRUNNLAG_URL}
 
-kodeverk.client.id = ${?KODEVERK_CLIENT_ID}
-kodeverk.resource.url = ${?KODEVERK_URL}
-applicationName = ${?NAIS_APP_NAME}
-
 funksjonsbrytere.unleash.applicationName = ${?NAIS_APP_NAME}
 funksjonsbrytere.unleash.host = ${?UNLEASH_SERVER_API_URL}
 funksjonsbrytere.unleash.token = ${?UNLEASH_SERVER_API_TOKEN}
